Combine all socket clients into RpcSocketClient (#2422)
* Combine all socket clients into RpcSocketClient

* Fix IPC sockets for windows

* Remove mention of node-ipc

* Remove less comprehensive test

* Remove node-ipc package
