+++ ironfish/package.json
@@ -38,7 +38,6 @@
     "lodash": "4.17.21",
     "node-datachannel": "0.3.6",
     "node-forge": "1.3.1",
-    "node-ipc": "9.1.3",
     "parse-json": "5.2.0",
     "sqlite": "4.0.23",
     "sqlite3": "5.0.4",
