+++ ironfish/src/rpc/clients/tlsClient.ts
@@ -2,12 +2,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import tls from 'tls'
+import { Assert } from '../../assert'
 import { ErrorUtils } from '../../utils'
 import { RpcConnectionRefusedError } from './errors'
 import { RpcTcpClient } from './tcpClient'
 
 export class RpcTlsClient extends RpcTcpClient {
   async connect(): Promise<void> {
+    this.logger.debug(`Connecting to ${this.describe()}`)
+
     return new Promise((resolve, reject): void => {
       const onSecureConnect = () => {
         client.off('secureConnection', onSecureConnect)
@@ -19,6 +22,7 @@ export class RpcTlsClient extends RpcTcpClient {
       const onError = (error: unknown) => {
         client.off('secureConnection', onSecureConnect)
         client.off('error', onError)
+
         if (ErrorUtils.isConnectRefusedError(error)) {
           reject(new RpcConnectionRefusedError())
         } else if (ErrorUtils.isNoEntityError(error)) {
@@ -32,9 +36,12 @@ export class RpcTlsClient extends RpcTcpClient {
         rejectUnauthorized: false,
       }
 
-      this.logger.debug(`Connecting to ${String(this.host)}:${String(this.port)}`)
-      const client = tls.connect(this.port, this.host, options, onSecureConnect)
+      Assert.isNotUndefined(this.connectTo.host)
+      Assert.isNotUndefined(this.connectTo.port)
+
+      const client = tls.connect(this.connectTo.port, this.connectTo.host, options)
       client.on('error', onError)
+      client.on('secureConnect', onSecureConnect)
       this.client = client
     })
   }
